Rename custom attributest wrapper
