+++ src/ggrc/assets/javascripts/apps/custom_attributes_wrap.js
